Added SyntaxWarning when using more that two ||
The algorithm used to replace || with Abs() cannot handle nested || (e.g. |a+|b||). This can probably be fixed (e.g. by checking whether each | is preceded by an operator, another | or a variable name to determine how they should be paired up), but for now the functionality is limited.